requirements.txt
Django==1.6.2
South==1.0.2
argparse==1.2.1
djangorestframework==2.3.13
six==1.9.0
wsgiref==0.1.2
